[ANDROSDK-2179] Rename BaseMultiUserDatabaseManager to MultiUserDatabaseManager
