+++ core/src/main/java/org/hisp/dhis/android/core/arch/d2/internal/D2DIComponent.kt
@@ -35,7 +35,7 @@ import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter
 import org.hisp.dhis.android.core.arch.storage.internal.CredentialsSecureStore
 import org.hisp.dhis.android.core.arch.storage.internal.UserIdInMemoryStore
 import org.hisp.dhis.android.core.category.internal.CategoryOptionStore
-import org.hisp.dhis.android.core.configuration.internal.BaseMultiUserDatabaseManager
+import org.hisp.dhis.android.core.configuration.internal.MultiUserDatabaseManager
 import org.hisp.dhis.android.core.configuration.internal.MultiUserDatabaseManagerForD2Manager
 import org.hisp.dhis.android.core.dataelement.internal.DataElementEndpointCallFactory
 import org.hisp.dhis.android.core.dataset.internal.DataSetEndpointCallFactory
@@ -113,7 +113,7 @@ internal class D2DIComponent(
     val interpreterSelector: InterpreterSelector,
 
     @get:VisibleForTesting
-    val multiUserDatabaseManager: BaseMultiUserDatabaseManager,
+    val multiUserDatabaseManager: MultiUserDatabaseManager,
 
     @get:VisibleForTesting
     val fileResourceDownloadCallHelper: FileResourceDownloadCallHelper,
