+++ core/src/main/java/org/hisp/dhis/android/core/configuration/internal/MultiUserDatabaseManager.kt
@@ -25,226 +25,105 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+
 package org.hisp.dhis.android.core.configuration.internal
 
-import android.content.Context
-import android.util.Log
 import org.hisp.dhis.android.core.arch.db.access.DatabaseExportMetadata
-import org.hisp.dhis.android.core.arch.db.access.DatabaseManager
-import org.hisp.dhis.android.core.arch.db.access.internal.BaseDatabaseExport
-import org.hisp.dhis.android.core.arch.helpers.FileResourceDirectoryHelper
 import org.hisp.dhis.android.core.arch.storage.internal.Credentials
-import org.hisp.dhis.android.core.util.CipherUtil
-import org.hisp.dhis.android.core.util.deleteIfExists
 
-@Suppress("TooManyFunctions")
-internal class MultiUserDatabaseManager(
-    private val context: Context,
-    private val databaseConfigurationSecureStore: DatabaseConfigurationInsecureStore,
-    private val configurationHelper: DatabaseConfigurationHelper,
-    private val databaseManager: DatabaseManager,
-    private val databaseExport: BaseDatabaseExport,
-) : BaseMultiUserDatabaseManager {
-    override suspend fun loadExistingChangingEncryptionIfRequiredOtherwiseCreateNew(
+/**
+ * Interface for managing multiple user databases in the DHIS2 Android SDK.
+ * This interface defines the operations required for supporting multiple users,
+ * including database creation, encryption, and import/export functionality.
+ */
+internal interface MultiUserDatabaseManager {
+
+    /**
+     * Loads an existing database, changes encryption if required, or creates a new database.
+     *
+     * @param serverUrl The server URL associated with the database
+     * @param username The username associated with the database
+     * @param encrypt Whether the database should be encrypted
+     */
+    suspend fun loadExistingChangingEncryptionIfRequiredOtherwiseCreateNew(
         serverUrl: String,
         username: String,
         encrypt: Boolean,
-    ) {
-        val existing = loadExistingChangingEncryptionIfRequired(
-            serverUrl,
-            username,
-            { encrypt },
-            true,
-        )
+    )
 
-        if (!existing) {
-            createNew(serverUrl, username, encrypt)
-        }
-    }
-
-    override suspend fun loadExistingKeepingEncryptionOtherwiseCreateNew(
+    /**
+     * Loads an existing database keeping its encryption status, or creates a new database.
+     *
+     * @param serverUrl The server URL associated with the database
+     * @param username The username associated with the database
+     * @param encrypt Whether the database should be encrypted if creating a new one
+     */
+    suspend fun loadExistingKeepingEncryptionOtherwiseCreateNew(
         serverUrl: String,
         username: String,
         encrypt: Boolean,
-    ) {
-        val existing = loadExistingChangingEncryptionIfRequired(
-            serverUrl,
-            username,
-            { obj: DatabaseAccount -> obj.encrypted() },
-            true,
-        )
+    )
 
-        if (!existing) {
-            createNew(serverUrl, username, encrypt)
-        }
-    }
+    /**
+     * Creates a new database for the specified user.
+     *
+     * @param serverUrl The server URL associated with the database
+     * @param username The username associated with the database
+     * @param encrypt Whether the database should be encrypted
+     */
+    fun createNew(serverUrl: String, username: String, encrypt: Boolean)
 
-    override fun createNew(serverUrl: String, username: String, encrypt: Boolean) {
-        removeExceedingAccounts()
-        val userConfiguration = addOrUpdateAccountInternal(serverUrl, username, encrypt)
-        databaseManager.createOrOpenDatabase(userConfiguration)
-    }
+    /**
+     * Creates a new database pending to be imported.
+     *
+     * @param metadata The metadata for the database to be imported
+     * @return The created database account
+     */
+    fun createNewPendingToImport(metadata: DatabaseExportMetadata): DatabaseAccount
 
-    override fun createNewPendingToImport(metadata: DatabaseExportMetadata): DatabaseAccount {
-        removeExceedingAccounts()
-        return addOrUpdateAccountInternal(
-            metadata.serverUrl,
-            metadata.username,
-            metadata.encrypted,
-            importStatus = DatabaseAccountImportStatus.PENDING_TO_IMPORT,
-        )
-    }
+    /**
+     * Changes encryption if required based on the provided credentials.
+     *
+     * @param credentials The user credentials
+     * @param encrypt Whether encryption should be enabled
+     */
+    suspend fun changeEncryptionIfRequired(credentials: Credentials, encrypt: Boolean)
 
-    override suspend fun changeEncryptionIfRequired(credentials: Credentials, encrypt: Boolean) {
-        loadExistingChangingEncryptionIfRequired(
-            credentials.serverUrl,
-            credentials.username,
-            { encrypt },
-            false,
-        )
-    }
+    /**
+     * Loads an existing database keeping its current encryption status.
+     *
+     * @param serverUrl The server URL associated with the database
+     * @param username The username associated with the database
+     * @return True if the database was loaded, false otherwise
+     */
+    suspend fun loadExistingKeepingEncryption(serverUrl: String, username: String): Boolean
 
-    override suspend fun loadExistingKeepingEncryption(serverUrl: String, username: String): Boolean {
-        return loadExistingChangingEncryptionIfRequired(
-            serverUrl,
-            username,
-            { obj: DatabaseAccount -> obj.encrypted() },
-            true,
-        )
-    }
+    /**
+     * Sets the maximum number of accounts that can be stored.
+     *
+     * @param maxAccounts The maximum number of accounts, or null for unlimited
+     */
+    fun setMaxAccounts(maxAccounts: Int?)
 
-    override fun setMaxAccounts(maxAccounts: Int?) {
-        if (maxAccounts != null && maxAccounts <= 0) {
-            throw IllegalArgumentException("MaxAccounts must be greater than 0")
-        } else {
-            val configuration = databaseConfigurationSecureStore.get()
-            val updatedConfiguration = (configuration?.toBuilder() ?: DatabasesConfiguration.builder())
-                .maxAccounts(maxAccounts)
-                .build()
-            databaseConfigurationSecureStore.set(updatedConfiguration)
-        }
-    }
+    /**
+     * Imports and loads a database.
+     *
+     * @param account The database account to import
+     * @param password The encryption password if applicable
+     */
+    suspend fun importAndLoadDb(account: DatabaseAccount, password: String)
 
-    @Suppress("TooGenericExceptionCaught")
-    override suspend fun importAndLoadDb(account: DatabaseAccount, password: String) {
-        val protectedDbPath = context.getDatabasePath(account.importDB()!!.protectedDbName())
-        val dbPath = context.getDatabasePath(account.databaseName())
-        val tempDbPath = context.filesDir.resolve("temp.db").also { it.deleteIfExists() }
-        try {
-            CipherUtil.extractEncryptedZipFile(protectedDbPath, tempDbPath, password)
-            protectedDbPath.deleteIfExists()
-
-            if (account.encrypted()) {
-                databaseExport.encryptAndCopyTo(account, sourceFile = tempDbPath, targetFile = dbPath)
-            } else {
-                tempDbPath.copyTo(dbPath)
-            }
-            databaseManager.createOrOpenDatabase(account)
-            val importedAccount = account.toBuilder()
-                .importDB(
-                    account.importDB()!!.toBuilder()
-                        .status(DatabaseAccountImportStatus.IMPORTED)
-                        .build(),
-                )
-                .build()
-            addOrUpdatedAccountInternal(importedAccount)
-        } catch (e: Exception) {
-            dbPath.deleteIfExists()
-            throw e
-        } finally {
-            tempDbPath.deleteIfExists()
-        }
-    }
-
-    private suspend fun loadExistingChangingEncryptionIfRequired(
-        serverUrl: String,
-        username: String,
-        encryptionExtractor: (config: DatabaseAccount) -> Boolean,
-        alsoOpenWhenEncryptionDoesntChange: Boolean,
-    ): Boolean {
-        val existingAccount = getAccount(serverUrl, username) ?: return false
-        val encrypt = encryptionExtractor(existingAccount)
-        changeEncryptionIfRequired(serverUrl, existingAccount, encrypt)
-        if (encrypt != existingAccount.encrypted() || alsoOpenWhenEncryptionDoesntChange) {
-            val updatedAccount = addOrUpdateAccountInternal(
-                serverUrl,
-                username,
-                encrypt,
-            )
-            databaseManager.createOrOpenDatabase(updatedAccount)
-        }
-        return true
-    }
-
-    private suspend fun changeEncryptionIfRequired(
-        serverUrl: String,
-        existingAccount: DatabaseAccount,
-        encrypt: Boolean,
-    ) {
-        if (encrypt != existingAccount.encrypted()) {
-            Log.w(
-                MultiUserDatabaseManager::class.java.name,
-                "Encryption value changed for " + existingAccount.username() + ": " + encrypt,
-            )
-            if (encrypt && !existingAccount.encrypted()) {
-                databaseExport.encrypt(serverUrl, existingAccount)
-            } else if (!encrypt && existingAccount.encrypted()) {
-                databaseExport.decrypt(serverUrl, existingAccount)
-            }
-            databaseManager.deleteDatabase(existingAccount.databaseName(), existingAccount.encrypted())
-        }
-    }
-
-    override fun getAccount(serverUrl: String, username: String): DatabaseAccount? {
-        val configuration = databaseConfigurationSecureStore.get()
-        return DatabaseConfigurationHelper.getAccount(configuration, serverUrl, username)
-    }
-
-    private fun addOrUpdateAccountInternal(
-        serverUrl: String,
-        username: String,
-        encrypt: Boolean,
-        importStatus: DatabaseAccountImportStatus? = null,
-    ): DatabaseAccount {
-        val updatedAccount = configurationHelper.addOrUpdateAccount(
-            databaseConfigurationSecureStore.get(),
-            serverUrl,
-            username,
-            encrypt,
-            null,
-            importStatus,
-        )
-        databaseConfigurationSecureStore.set(updatedAccount)
-        return DatabaseConfigurationHelper.getLoggedAccount(updatedAccount, username, serverUrl)
-    }
-
-    private fun addOrUpdatedAccountInternal(account: DatabaseAccount) {
-        val updatedAccount = configurationHelper.addOrUpdateAccount(
-            databaseConfigurationSecureStore.get(),
-            account,
-        )
-        databaseConfigurationSecureStore.set(updatedAccount)
-    }
-
-    private fun removeExceedingAccounts() {
-        val configuration = databaseConfigurationSecureStore.get()
-
-        configuration?.maxAccounts()?.let { maxAccounts ->
-            val exceedingAccounts = DatabaseConfigurationHelper
-                .getOldestAccounts(configuration.accounts(), maxAccounts - 1)
-
-            val updatedConfiguration =
-                DatabaseConfigurationHelper.removeAccount(configuration, exceedingAccounts)
-
-            databaseConfigurationSecureStore.set(updatedConfiguration)
-            exceedingAccounts.forEach {
-                FileResourceDirectoryHelper.deleteFileResourceDirectories(context, it)
-                databaseManager.deleteDatabase(it.databaseName(), it.encrypted())
-            }
-        }
-    }
+    /**
+     * Gets an account by server URL and username.
+     *
+     * @param serverUrl The server URL associated with the account
+     * @param username The username associated with the account
+     * @return The database account, or null if not found
+     */
+    fun getAccount(serverUrl: String, username: String): DatabaseAccount?
 
     companion object {
+        const val DefaultMaxAccounts = 1
         internal val DefaultTestMaxAccounts = null
     }
 }
